FU-106 refactor: separate schedule controller
