+++ src/components/calender/schedule.css.ts
@@ -1,0 +1,12 @@
+import { styleVariants } from "@vanilla-extract/css";
+
+export const ScheduleCalenderStyles = styleVariants({
+  container: {
+    paddingTop: 20,
+    paddingBottom: 20,
+    height: 540,
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "space-between",
+  },
+});
