+++ src/containers/customer/reservation/submit/schedule/schedule-select.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
-import { DatePicker, DateValue } from "@mantine/dates";
+import { DateValue } from "@mantine/dates";
 import { useFormContext, useFieldArray } from "react-hook-form";
 import "dayjs/locale/ko";
 import { reservation } from "product-types";
-import CustomedTimeInput from "@/components/inputs/time-input";
+import ScheduleCalender from "@/components/calender/schedule-calender";
 import ScheduleLayout from "../schedule-layout";
 import submitStyles from "../submit.css";
 
@@ -20,40 +20,14 @@ const ScheduleSelect = () => {
     name: "schedules",
   });
 
-  function handleSelectNewDate(newDate: DateValue) {
-    setValueOnModal((prev) => {
-      return { ...prev, date: newDate };
-    });
-  }
-
-  function handleSelectNewTime(newTime: Date, target: "startTime" | "endTime") {
-    setValueOnModal((prev) => {
-      return { ...prev, target: newTime };
-    });
-  }
-
   return (
     <ScheduleLayout>
       <span className={submitStyles.title}>후보 일정 추가하기</span>
-      <div className={submitStyles.calenderWrapper}>
-        <DatePicker
-          locale="ko"
-          size="lg"
-          firstDayOfWeek={0}
-          allowDeselect
-          onChange={handleSelectNewDate}
-        />
+      <div>
+        <div className={submitStyles.calenderWrapper}>
+          <ScheduleCalender value={valueOnModal} setValue={setValueOnModal} />
+        </div>
       </div>
-      <CustomedTimeInput
-        title="시작"
-        date={valueOnModal.date ? valueOnModal.date : undefined}
-        onChangeTime={(time) => handleSelectNewTime(time, "startTime")}
-      />
-      <CustomedTimeInput
-        title="종료"
-        date={valueOnModal.date ? valueOnModal.date : undefined}
-        onChangeTime={(time) => handleSelectNewTime(time, "endTime")}
-      />
     </ScheduleLayout>
   );
 };
